LayoutTests/imported/w3c/web-platform-tests/screen-wake-lock/wakelock-document-hidden.https-expected.txt


PASS navigator.wakeLock.request('screen') fails when the document is hidden
PASS Screen wake locks are released when the document the page is hidden
